DeepLearning/requirements.txt
absl-py==0.9.0
aiohttp==3.7.3
astor==0.8.1
astroid==2.2.5
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
distlib==0.3.1
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.2
lxml==4.5.1
Markdown==3.2.2
matplotlib @ file:///C:/ci/matplotlib-base_1592846084747/work
mccabe==0.6.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
Pillow==7.1.2
pipenv==2020.6.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==4.6.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
tensorboard==2.1.1
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
tensornets==0.4.6
termcolor==1.1.0
tornado==6.0.4
tqdm==4.46.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
yarl==1.6.3
zipp==3.1.0
